Reviews.py
The code in this file has been updated in order to create the review model. Now, we have the id for the movie, a char field for the length of the comments, a field for the date and time the review was created, a function that deletes the review if the movie associated with the review is deleted, and a method that returns a string representation of the review.